pwa/requirements.txt
flask==3.0.3
gunicorn==22.0.0
gspread==6.1.2
oauth2client==4.1.3
python-dateutil==2.9.0
tzdata==2024.1
supabase==2.8.1
pyjwt==2.9.0
